+++ lambda_erp/buying/purchase_order.py
@@ -14,7 +14,10 @@ from lambda_erp.model import Document
 from lambda_erp.utils import _dict, flt, getdate, nowdate
 from lambda_erp.database import get_db
 from lambda_erp.controllers.taxes_and_totals import calculate_taxes_and_totals
-from lambda_erp.controllers.defaults import set_default_currency
+from lambda_erp.controllers.defaults import (
+    set_default_currency, apply_external_source_defaults,
+    derived_pricing_fields, derived_line_pricing_fields,
+)
 from lambda_erp.controllers.item_price import set_item_defaults
 from lambda_erp.exceptions import ValidationError
 
@@ -50,6 +53,7 @@ class PurchaseOrder(Document):
         if not self.transaction_date:
             self.transaction_date = nowdate()
 
+        apply_external_source_defaults(self)
         self._set_supplier_name()
         set_default_currency(self, "Supplier", "supplier")
         self._set_item_defaults()
@@ -91,6 +95,7 @@ def make_purchase_invoice(purchase_order_name):
         raise ValidationError("Purchase Order must be submitted before creating Purchase Invoice")
 
     pi = PurchaseInvoice(
+        **derived_pricing_fields(po, is_return=False),
         supplier=po.supplier,
         supplier_name=po.supplier_name,
         company=po.company,
@@ -106,6 +111,7 @@ def make_purchase_invoice(purchase_order_name):
             continue
 
         pi.append("items", _dict(
+            **derived_line_pricing_fields(item),
             item_code=item.get("item_code"),
             item_name=item.get("item_name"),
             description=item.get("description"),